+++ lib/redis/connection/synchrony.rb
@@ -6,8 +6,8 @@ require_relative "../errors"
 require "em-synchrony"
 require "hiredis/reader"
 
-Kernel.warn(
-  "The redis synchrony driver is deprecated and will be removed in redis-rb 5.0. " \
+::Redis.deprecate!(
+  "The redis synchrony driver is deprecated and will be removed in redis-rb 5.0.0. " \
   "We're looking for people to maintain it as a separate gem, see https://github.com/redis/redis-rb/issues/915"
 )