+++ test/connection_test.rb
@@ -30,7 +30,6 @@ class TestConnection < Minitest::Test
   def test_connection_with_wrong_user_and_password
     target_version "6.0" do
       with_default_user_password do |_username, password|
-        Kernel.expects(:warn).once
         redis = Redis.new(OPTIONS.merge(username: "does-not-exist", password: password))
         assert_equal "PONG", redis.ping
       end