+++ test/pipelining_commands_test.rb
@@ -6,9 +6,9 @@ class TestPipeliningCommands < Minitest::Test
   include Helper::Client
 
   def test_bulk_commands
-    r.pipelined do
-      r.lpush "foo", "s1"
-      r.lpush "foo", "s2"
+    r.pipelined do |p|
+      p.lpush "foo", "s1"
+      p.lpush "foo", "s2"
     end
 
     assert_equal 2, r.llen("foo")
@@ -17,9 +17,9 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_multi_bulk_commands
-    r.pipelined do
-      r.mset("foo", "s1", "bar", "s2")
-      r.mset("baz", "s3", "qux", "s4")
+    r.pipelined do |p|
+      p.mset("foo", "s1", "bar", "s2")
+      p.mset("baz", "s3", "qux", "s4")
     end
 
     assert_equal "s1", r.get("foo")
@@ -29,10 +29,10 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_bulk_and_multi_bulk_commands_mixed
-    r.pipelined do
-      r.lpush "foo", "s1"
-      r.lpush "foo", "s2"
-      r.mset("baz", "s3", "qux", "s4")
+    r.pipelined do |p|
+      p.lpush "foo", "s1"
+      p.lpush "foo", "s2"
+      p.mset("baz", "s3", "qux", "s4")
     end
 
     assert_equal 2, r.llen("foo")
@@ -43,10 +43,10 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_multi_bulk_and_bulk_commands_mixed
-    r.pipelined do
-      r.mset("baz", "s3", "qux", "s4")
-      r.lpush "foo", "s1"
-      r.lpush "foo", "s2"
+    r.pipelined do |p|
+      p.mset("baz", "s3", "qux", "s4")
+      p.lpush "foo", "s1"
+      p.lpush "foo", "s2"
     end
 
     assert_equal 2, r.llen("foo")
@@ -64,19 +64,19 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_returning_the_result_of_a_pipeline
-    result = r.pipelined do
-      r.set "foo", "bar"
-      r.get "foo"
-      r.get "bar"
+    result = r.pipelined do |p|
+      p.set "foo", "bar"
+      p.get "foo"
+      p.get "bar"
     end
 
     assert_equal ["OK", "bar", nil], result
   end
 
   def test_assignment_of_results_inside_the_block
-    r.pipelined do
-      @first = r.sadd("foo", 1)
-      @second = r.sadd("foo", 1)
+    r.pipelined do |p|
+      @first = p.sadd("foo", 1)
+      @second = p.sadd("foo", 1)
     end
 
     assert_equal true, @first.value
@@ -87,10 +87,10 @@ class TestPipeliningCommands < Minitest::Test
   # it doesn't make a lot of sense.
   def test_assignment_of_results_inside_the_block_with_errors
     assert_raises(Redis::CommandError) do
-      r.pipelined do
-        r.doesnt_exist
-        @first = r.sadd("foo", 1)
-        @second = r.sadd("foo", 1)
+      r.pipelined do |p|
+        p.doesnt_exist
+        @first = p.sadd("foo", 1)
+        @second = p.sadd("foo", 1)
       end
     end
 
@@ -99,11 +99,11 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_assignment_of_results_inside_a_nested_block
-    r.pipelined do
-      @first = r.sadd("foo", 1)
+    r.pipelined do |p|
+      @first = p.sadd("foo", 1)
 
-      r.pipelined do
-        @second = r.sadd("foo", 1)
+      r.pipelined do |p2|
+        @second = p2.sadd("foo", 1)
       end
     end
 
@@ -112,32 +112,32 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_futures_raise_when_confused_with_something_else
-    r.pipelined do
-      @result = r.sadd("foo", 1)
+    r.pipelined do |p|
+      @result = p.sadd("foo", 1)
     end
 
     assert_raises(NoMethodError) { @result.to_s }
   end
 
   def test_futures_raise_when_trying_to_access_their_values_too_early
-    r.pipelined do
+    r.pipelined do |p|
       assert_raises(Redis::FutureNotReady) do
-        r.sadd("foo", 1).value
+        p.sadd("foo", 1).value
       end
     end
   end
 
   def test_futures_raise_when_command_errors_and_needs_transformation
     assert_raises(Redis::CommandError) do
-      r.pipelined do
-        @result = r.zrange("a", "b", 5, with_scores: true)
+      r.pipelined do |p|
+        @result = p.zrange("a", "b", 5, with_scores: true)
       end
     end
   end
 
   def test_futures_warn_when_tested_for_equality
-    r.pipelined do
-      @result = r.sadd("foo", 1)
+    r.pipelined do |p|
+      @result = p.sadd("foo", 1)
     end
 
     Redis.expects(:deprecate!).once
@@ -145,8 +145,8 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_futures_can_be_identified
-    r.pipelined do
-      @result = r.sadd("foo", 1)
+    r.pipelined do |p|
+      @result = p.sadd("foo", 1)
     end
 
     assert_equal true, @result.is_a?(Redis::Future)
@@ -162,10 +162,10 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_nesting_pipeline_blocks
-    r.pipelined do
-      r.set("foo", "s1")
-      r.pipelined do
-        r.set("bar", "s2")
+    r.pipelined do |p|
+      p.set("foo", "s1")
+      p.pipelined do |p2|
+        p2.set("bar", "s2")
       end
     end
 
@@ -174,16 +174,16 @@ class TestPipeliningCommands < Minitest::Test
   end
 
   def test_info_in_a_pipeline_returns_hash
-    result = r.pipelined do
-      r.info
+    result = r.pipelined do |p|
+      p.info
     end
 
     assert result.first.is_a?(Hash)
   end
 
   def test_config_get_in_a_pipeline_returns_hash
-    result = r.pipelined do
-      r.config(:get, "*")
+    result = r.pipelined do |p|
+      p.config(:get, "*")
     end
 
     assert result.first.is_a?(Hash)
@@ -191,8 +191,8 @@ class TestPipeliningCommands < Minitest::Test
 
   def test_hgetall_in_a_pipeline_returns_hash
     r.hmset("hash", "field", "value")
-    result = r.pipelined do
-      r.hgetall("hash")
+    result = r.pipelined do |p|
+      p.hgetall("hash")
     end
 
     assert_equal result.first, { "field" => "value" }
@@ -200,8 +200,8 @@ class TestPipeliningCommands < Minitest::Test
 
   def test_keys_in_a_pipeline
     r.set("key", "value")
-    result = r.pipelined do
-      r.keys("*")
+    result = r.pipelined do |p|
+      p.keys("*")
     end
 
     assert_equal ["key"], result.first