+++ lib/redis.rb
@@ -28,11 +28,15 @@ class Redis
       ::Kernel.warn(message) unless silence_deprecations
     end
 
-    attr_writer :current
-  end
+    def current
+      deprecate!("`Redis.current=` is deprecated and will be removed in 5.0. (called from: #{caller(1, 1).first})")
+      @current ||= Redis.new
+    end
 
-  def self.current
-    @current ||= Redis.new
+    def current=(redis)
+      deprecate!("`Redis.current=` is deprecated and will be removed in 5.0. (called from: #{caller(1, 1).first})")
+      @current = redis
+    end
   end
 
   include Commands