+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* Fix `zpopmax` and `zpopmin` when called inside a pipeline. See #1055.
+* Deprecate `Redis.current`.
 * Deprecate calling commands on `Redis` inside `Redis#pipelined`. See #1059.
   ```ruby
   redis.pipelined do
@@ -16,6 +16,9 @@
   end
   ```
 * Deprecate `Redis#queue` and `Redis#commit`. See #1059.
+
+* Fix `zpopmax` and `zpopmin` when called inside a pipeline. See #1055.
+
 * Add `Redis.silence_deprecations=` to turn off deprecation warnings.
   If you don't wish to see warnings yet, you can set `Redis.silence_deprecations = false`.
   It is however heavily recommended to fix them instead when possible.